+++ lib/braintree/transaction_gateway.rb
@@ -223,7 +223,8 @@ module Braintree
           {:paypal => [:custom_field, :payee_id, :payee_email, :description, {:supplementary_data => :_any_key_}]},
           {:three_d_secure => [:required]},
           {:amex_rewards => [:request_id, :points, :currency_amount, :currency_iso_code]},
-          {:venmo => [:profile_id]}
+          {:venmo => [:profile_id]},
+          {:credit_card => [:account_type]},
         ]
         },
         {:external_vault => [