+++ spec/spec_helper.rb
@@ -43,6 +43,7 @@ unless defined?(SPEC_HELPER_LOADED)
     AnotherUsBankMerchantAccountId = "another_us_bank_merchant_account"
     IdealMerchantAccountId = "ideal_merchant_account"
     AdyenMerchantAccountId = "adyen_ma"
+    HiperBRLMerchantAccountId = "hiper_brl"
 
     TrialPlan = {
       :description => "Plan for integration tests -- with trial",